requirements.txt
langchain
langgraph
langchain-openai
langchain-deepseek
langchain-ollama
python-dotenv
sentence-transformers
scikit-learn
numpy